App now adds a new Expense Item to the list dynamically
Found the bug that was causing the code to fail when adding an expense item to the list.

It is crucial that an arrow function is returned to the setter updating function setExpense!!!. The arrow function must return a concatenated array with the new expense item plus the previous expenses array. The previous expenses array is automatically passed to the method, you just have to name it.